feat(manager-pci-common): refactor modal and rename banners components
