+++ packages/manager/modules/manager-pci-common/src/components/banner/discovery-banner/PciDiscoveryBanner.tsx
