--- packages/manager/modules/manager-pci-common/src/components/deletion-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './DeletionModal.component';
