+++ packages/manager/modules/manager-pci-common/src/components/modal/index.ts
@@ -1,1 +1,2 @@
 export * from './PciModal.component';
+export * from './deletion-modal/DeletionModal.component';
