+++ packages/manager/modules/manager-pci-common/src/index.ts
@@ -3,7 +3,6 @@ export * from './api/hook';
 export * from './components/banner';
 export * from './components/region-selector/RegionSelector.component';
 export * from './components/region-selector/RegionSummary.component';
-export * from './components/deletion-modal';
 export * from './components/modal';
 export * from './components/flavor-selector';
 export * from './components/region-selector/RegionSelector.component';
